[Solved for iOS 14+] After implement Combine operators, the tests are passing, but if I try to run the app on the simulator or in a real device, the feed is not loaded anymore.
This seems like a Combine issue introduced on iOS 14. It works fine on iOS 13 - I can only replicate it on iOS 14.

The `RemoteFeedLoader` is being deallocated before it completes the operation because the Combine publisher is not holding a reference to it anymore. To solve it, you can create a property to hold a reference to the `RemoteFeedLoader` instance in the `SceneDelegate`.