Fix for iOS 15 cell lifecycle behavior change
Capture cell reference and load/reload cell resources on `willDisplayCell` as a fix for iOS 15 cell prefetching behavior change. On iOS 15+, for performance reasons, the table view data source may not recreate a cell using the `cellForRow` method if there's a cached cell for the given IndexPath. In this case, it'll just call `willDisplayCell`. However, we release a reference of the cell and cancel requests on `didEndDisplayingCell`. So, since `cellForRow` may not be called, we need to implement `willDisplayCell` to know when the cached cell is becoming visible again to recapture a reference of the cell and load/reload any resource needed for the cell.